Preguntas, GexResponse generic, etc

+++ Gex.NetCore/Repository/Interface/IComisionRepository.cs
@@ -7,12 +7,12 @@ namespace Gex.Repository.Interface;
 public interface IComisionRepository
 {
     Task<ICollection<Comision>> GetComisionsAsync();
-    Task<Comision> GetComisionAsync(int id);
+    Task<Comision> GetComisionAsync(long id);
     Task<Comision> GetComisionAsync(string nombre);
-    Task<bool> ExistsComisionAsync(int id);
+    Task<bool> ExistsComisionAsync(long id);
     Task<bool> CreateComisionAsync(Comision comision);
     Task<bool> DeleteComisionAsync(Comision comision);
-    Task<bool> DeleteComisionAsync(int id);
+    Task<bool> DeleteComisionAsync(long id);
     Task<bool> UpdateComisionAsync(Comision comision);
     Task<bool> ExistsComisionAsync(string nombre);
